chat-history/pyproject.toml
[project]
name = "chat-history"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "robertomitugno", email = "roberto.mitugno@studio.unibo.it" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "motor>=3.7.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]

[project.scripts]
chat-history = "chat_history:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
]
